Move new PublisherBuilder logic
